requirements.txt
blinker==1.6.2
Bootstrap-Flask==2.2.0
click==8.1.4
Flask==2.3.2
Flask-WTF==1.1.1
gunicorn==20.1.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Pillow==10.0.0
Werkzeug==2.3.6
WTForms==3.0.1
zipp==3.16.0
